Unregister all metrics upon error
If data fetching from ZTE device fails for any reason the metrics should
not be exported. Without this unregistering exporter continued to show
historical data which surely was no longer relevant.

Ideally it should respond with 404 or 5xx (inability to fetch expected data
is a server-side error after all) in order for "up==0" to trigger an alert
but this improvement can come later. For now it's better to have a gap in
Prometheus series than false data.